+++ blacklight_security/scanners/aws/__init__.py
@@ -2,6 +2,7 @@
 
 from .cloudtrail import CloudTrailScanner
 from .ec2 import EC2Scanner
+from .guardduty import GuardDutyScanner
 from .iam import IAMScanner
 from .lambda_functions import LambdaScanner
 from .rds import RDSScanner
@@ -10,6 +11,7 @@ from .s3 import S3Scanner
 __all__ = [
     "CloudTrailScanner",
     "EC2Scanner",
+    "GuardDutyScanner",
     "IAMScanner",
     "LambdaScanner",
     "RDSScanner",